web3kairo-bot/requirements.txt
python-telegram-bot>=20.0
feedparser>=6.0
requests>=2.28
